HOTFIX moving time treatment to different file
