+++ src/utils/scheduler.js
@@ -2,31 +2,9 @@ const schedule = require('node-schedule');
 const mongoose = require('mongoose');
 const axios = require('axios');
 const UserSchema = require('../db/userSchema');
+const TreatTime = require('./treatTime');
 
 const UserModel = mongoose.model('UserModel', UserSchema);
-
-function getDateTime() {
-  const today = new Date();
-  let time = today.getHours() - 3;
-  let weekDay = today.getDay();
-  if (time <= 0) {
-    if (time === -1) {
-      time = 23;
-      this.date = `${today.getDate() - 1}/${today.getMonth() + 1}/${today.getFullYear()}`;
-      weekDay -= 1;
-    } else if (time === -2) {
-      time = 22;
-      this.date = `${today.getDate() - 1}/${today.getMonth() + 1}/${today.getFullYear()}`;
-      weekDay -= 1;
-    } else if (time === -3) {
-      time = 21;
-      this.date = `${today.getDate() - 1}/${today.getMonth() + 1}/${today.getFullYear()}`;
-      weekDay -= 1;
-    }
-  }
-
-  return (weekDay);
-}
 
 function getDailyNotifications(weekDay) {
   const notificationArray = new Array([]);
@@ -64,7 +42,7 @@ async function makeSchedule(notification) {
 }
 
 function notificationSchedule() {
-  const weekDay = getDateTime();
+  const weekDay = TreatTime.getDateTime();
   getDailyNotifications(weekDay).then((notificationArray) => {
     for (let i = 0; i < notificationArray.length; i += 1) {
       makeSchedule(notificationArray[i]);
@@ -83,6 +61,5 @@ module.exports = {
   postNotification,
   makeSchedule,
   getDailyNotifications,
-  getDateTime,
   notificationSchedule,
 };
